+++ src/Subscriber/ConfigurationSubscriber.php
@@ -6,6 +6,7 @@ namespace Omikron\FactFinder\Shopware6\Subscriber;
 
 use Exception;
 use Omikron\FactFinder\Shopware6\Config\Communication;
+use Omikron\FactFinder\Shopware6\Config\ExtensionConfig;
 use Shopware\Core\Framework\Event\ShopwareSalesChannelEvent;
 use Shopware\Core\Framework\Struct\ArrayEntity;
 use Shopware\Core\Framework\Struct\Struct;
@@ -16,17 +17,20 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 class ConfigurationSubscriber implements EventSubscriberInterface
 {
     private Communication $config;
+    private ExtensionConfig $extensionConfig;
     private array $fieldRoles;
     private array $communicationParameters;
     private array $addParams;
 
     public function __construct(
         Communication $config,
+        ExtensionConfig $extensionConfig,
         array $fieldRoles,
         array $communicationParameters,
         array $configurationAddParams = []
     ) {
         $this->config                  = $config;
+        $this->extensionConfig         = $extensionConfig;
         $this->fieldRoles              = $fieldRoles;
         $this->communicationParameters = $communicationParameters;
         $this->addParams               = $configurationAddParams;
@@ -67,7 +71,8 @@ class ConfigurationSubscriber implements EventSubscriberInterface
             $page->addExtension('factfinder', new ArrayEntity([
                 'field_roles'      => $this->config->getFieldRoles($salesChannelId) ?: $this->fieldRoles,
                 'communication'    => $communication + $this->communicationParameters,
-                'trackingSettings' => $this->config->getTrackingSettings(),
+                'trackingSettings' => $this->extensionConfig->getTrackingSettings(),
+                'redirectMapping'  => (string) $this->extensionConfig->getRedirectMapping(),
                 'searchImmediate'  => $this->isSearchImmediate($event) ? 'true' : 'false',
                 'userId'           => $customer ? $customer->getId() : null,
                 'ssr'              => $this->config->isSsrActive(),